+++ docs/general/changelog.md
@@ -1,6 +1,19 @@
 # Module Changelog
 * ***Version 1.0.0***
   * Release module
-* ***Version 1.0.2***
+* ***Version 1.0.1 - 1.0.2***
   * JSDoc Fixes
+  * Changing data returned by `roleAdded` & `roleRemoved` events
   * Fix methods: `<Utils>.parseEmoji`, `<Utils>.fetchPartials`, `<ReactionsRoles>.edit`
+  * Update module types
+  * Release module documentation [drr-website](https://drr.js.org)
+* ***Version 1.0.3 - 1.0.4***
+  * Fix `Unknown emoji with content` error (`<Utils>.parseEmoji` method)
+  * Added methods `<ReactionsRole>.get` & `<ReactionsRoles>.list`
+  * Update module types
+  * Update module examples in docs
+  * Update all dependencies
+* ***Version 1.0.5***
+  * Update versions of all package dependencies
+  * Minor edits to JSDoc
+  * Fix some bugs